requirements.txt
cmake
djangorestframework
python-dateutil
Pyjwt
Django==2.2.4
django-cors-headers
nose
pymongo==3.9.0
six==1.15.0
flake8
flake8-import-order
openpyxl==2.5.6
django-crontab==0.7.1
Pillow==8.4.0
numpy==1.19.3
Click==7.0
dnspython==2.0.0
setuptools==41.0.0
psutil==5.8.0
unidecode==1.2.0
opencv-python==4.4.0.44
importlib-metadata==4.2.0
tensorflow==2.6.0
keras==2.6.0
protobuf==3.20.2
seaborn==0.13.2
PyYAML>=5.3.1
requests>=2.23.0
scipy>=1.4.1
torch>=1.7.0,!=1.12.0
torchvision>=0.8.1,!=0.13.0